fix: extract shared types from server/ to src/
- server 폴더 하위에 있던 client 와 server 가 공유하는 타입을 src 폴더 하위로 이동
- client 에서 server 폴더를 보게 되면서 absolute path 사용이 안되는 문제가 발생해서 변경하였음
