+++ src/server/api/add-to-cart.ts
@@ -1,14 +1,10 @@
 import { CartCreateInput, CartUpdateInput } from '@prisma/client'
 import express, { Request, Response } from 'express'
 import { body } from 'express-validator'
-import { STATUS_CODE, ERROR_MSG, CONSTRAINT } from '~/../constants'
+import { CONSTRAINT, ERROR_MSG, STATUS_CODE } from '~/../constants'
+import { AddToCartRequestBody } from '~/../types/api'
 import { requestValidator } from '~/middlewares'
 import { prisma } from '../utils/prisma'
-
-type AddToCartRequestBody = {
-  productId: number
-  quantity: number
-}
 
 const addToCartRouter = express.Router()
