+++ src/server/api/get-jjims.ts
@@ -1,12 +1,9 @@
-import { Jjim } from '@prisma/client'
 import express, { Request, Response } from 'express'
 import { ERROR_MSG, STATUS_CODE } from '~/../constants'
-import { ErrorResponse } from '~/types/res'
+import { GetJjimsApiResponse } from '~/../types/api'
 import { prisma } from '../utils/prisma'
 
 const getJjimsRouter = express.Router()
-
-export type GetJjimsApiResponse = Jjim[] | ErrorResponse
 
 getJjimsRouter.get('/jjims', async (req: Request, res: Response<GetJjimsApiResponse>) => {
   const userId = req.auth?.userId
