+++ src/server/api/set-default-address.ts
@@ -1,14 +1,11 @@
 import express, { Request, Response } from 'express'
-import { prisma } from '../utils/prisma'
-import { STATUS_CODE, ERROR_MSG } from '~/../constants'
 import { body } from 'express-validator'
+import { ERROR_MSG, STATUS_CODE } from '~/../constants'
+import { SetDefaultAddressApiRequestBody } from '~/../types/api'
 import { requestValidator } from '~/middlewares'
+import { prisma } from '../utils/prisma'
 
 const setDefaultAddressRouter = express.Router()
-
-export type SetDefaultAddressApiRequestBody = {
-  defaultAddressId: number
-}
 
 setDefaultAddressRouter.patch(
   '/set-default-address',
