--- src/server/types/res.d.ts
@@ -1,3 +1,0 @@
-export type ErrorResponse = {
-  message: string
-}
